Утилитарные функии BurgerConstructor перенесены в файл утилит компонента
