+++ src/components/BurgerConstructor/BurgerConstructor.utils.js
@@ -12,3 +12,41 @@ export const orderIdReducer = (_, action) => {
       throw new Error(`Неверный тип action: ${action.type}`);
   }
 };
+
+const getUniqueSelectedIngredientsByIds = (ingredientIds, ingredientsData) => {
+  return ingredientIds.map((id, index) => {
+    const found = ingredientsData.find((item) => item._id === id);
+    return {
+      ...found,
+      uniqueId: `${id}_${index}`,
+    };
+  });
+};
+
+const splitToBunAndInner = (selectedIngredientsData) => {
+  const bun = selectedIngredientsData.find((item) => item.type === "bun");
+  const innerIngredients = selectedIngredientsData.filter(
+    (item) => item.type !== "bun"
+  );
+
+  return { bun, innerIngredients };
+};
+
+export const getSplittedIngredientsData = (
+  selectedIngredientsIds,
+  ingredientsData
+) => {
+  const selectedIngredientsAll = getUniqueSelectedIngredientsByIds(
+    selectedIngredientsIds,
+    ingredientsData
+  );
+
+  return splitToBunAndInner(selectedIngredientsAll);
+};
+
+export const getBurgerTotalPrice = (bun, innerIngredients) => {
+  return (
+    bun.price * 2 +
+    innerIngredients.reduce((total, curr) => total + curr.price, 0)
+  );
+};
